+++ configurations/walmart/es5-node.js
@@ -1,8 +1,12 @@
-'use strict';
+"use strict";
 
-var _ = require('lodash');
+var _ = require("lodash");
 
 module.exports = _.merge({},
-  require('./es5.js'),
-  require('../../rules/eslint/node/walmart.js')
+  require("./es5.js"),
+  require("../../rules/eslint/node/walmart.js"), {
+    "rules": {
+      "strict": [2, "global"]
+    }
+  }
 );