+++ configurations/walmart/es6-node.js
@@ -1,8 +1,8 @@
-'use strict';
+"use strict";
 
-var _ = require('lodash');
+var _ = require("lodash");
 
 module.exports = _.merge({},
-  require('./es6.js'),
-  require('../../rules/eslint/node/walmart.js')
+  require("./es6.js"),
+  require("../../rules/eslint/node/walmart.js")
 );